type error fixed. build should be ok again
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@6632 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/components/DVB/DvbTeleText.cpp
@@ -109,7 +109,7 @@ void
 DvbTeleText::set_filter(){
     struct dmx_pes_filter_params f;
     
-    f.pid = (uint16_t) _myPid;
+    f.pid = _myPid;
     f.input = DMX_IN_FRONTEND;
     f.output = DMX_OUT_TAP;
     f.pes_type = DMX_PES_OTHER; /* DMX_PES_TELETEXT if you want vbi insertion */
